converting value serializer basic tests to data oriented ones

+++ Fracture.Net/Serialization/ShortSerializer.cs
@@ -35,13 +35,13 @@ namespace Fracture.Net.Serialization
         /// <summary>
         /// Returns size of int16, should always be 2-bytes.
         /// </summary>
-        public override int GetSizeFromBuffer(byte[] buffer, int offset)
+        public override ushort GetSizeFromBuffer(byte[] buffer, int offset)
             => sizeof(short);
         
         /// <summary>
         /// Returns size of int16, should always be 2-bytes.
         /// </summary>
-        public override int GetSizeFromValue(object value)
+        public override ushort GetSizeFromValue(object value)
             => sizeof(short);
     }
